Move config-related things into the Txgh::Config namespace
