+++ .gitignore
@@ -23,4 +23,4 @@ Gemfile.lock
 config/txgh_config.rb
 .env
 .byebug_history
-
+.ruby-version
