+++ lib/txgh.rb
@@ -9,7 +9,6 @@ module Txgh
   autoload :GithubRequestAuth,    'txgh/github_request_auth'
   autoload :Handlers,             'txgh/handlers'
   autoload :Hooks,                'txgh/app'
-  autoload :KeyManager,           'txgh/key_manager'
   autoload :ParseConfig,          'txgh/parse_config'
   autoload :ResourceCommitter,    'txgh/resource_committer'
   autoload :ResourceUpdater,      'txgh/resource_updater'
@@ -19,7 +18,6 @@ module Txgh
   autoload :TransifexRequestAuth, 'txgh/transifex_request_auth'
   autoload :Triggers,             'txgh/app'
   autoload :TxBranchResource,     'txgh/tx_branch_resource'
-  autoload :TxConfig,             'txgh/tx_config'
   autoload :TxLogger,             'txgh/tx_logger'
   autoload :TxResource,           'txgh/tx_resource'
   autoload :Utils,                'txgh/utils'
