+++ spec/config/key_manager_spec.rb
@@ -2,6 +2,7 @@ require 'spec_helper'
 require 'helpers/standard_txgh_setup'
 
 include Txgh
+include Txgh::Config
 
 describe KeyManager do
   include StandardTxghSetup
